% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMIndex_Script.R
\name{createBWT}
\alias{createBWT}
\title{\strong{Creates a Burrows-Wheeler Transform of a DNA string}}
\usage{
createBWT(string)
}
\arguments{
\item{string}{a character string representing a DNA sequence}
}
\value{
a vector of 5 strings.
The first 4 elements represent the count of A's, C's, G's, and T's.
The last element is the BWT of the DNA string.
}
\description{
This function takes a DNA string as input, containing the bases A,C,G and T.
Then creates a Burrows-Wheeler Transformation (BWT) of that string through
the usage of a suffix array.
The BWT is the last column of the Burrows-Wheeler Matrix (BWM).
It gives in output a vector of 5 character containing the BWT
(as the last element) and 4 strings representing a compressed version of the
first column of the Burrows_Wheeler Matrix (BWM).
These strings correspond to the counts of the bases A, C, G, and T
(in that order) in the first column of the BWM, respectively.
}
\examples{
DNA_string <- "ACCTGGAC"
bwt <- createBWT(DNA_string)
print(bwt)

}
\seealso{
\code{\link{create_SA}} for generating the suffix array used in this
function.
}
