% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMIndex_Script.R
\name{FM_index}
\alias{FM_index}
\title{\strong{Create a compressed FM-index from a DNA Sequence}}
\usage{
FM_index(fasta_file, step_tally = 32, k = 32, output = "./FM_index_Results")
}
\arguments{
\item{fasta_file}{A string representing the path to a FASTA file containing
a DNA sequence.}

\item{step_tally}{An integer specifying the step size for generating the
sparse tally matrix.
Default is 32.}

\item{k}{An integer specifying the step size for selecting elements from the
suffix array to create the SSA.
Default is 32.}

\item{output}{A string specifying the folder that will contain the output
files (BWT, sparse tally, SSA) should be saved.
If it does not exist, will be created.
Default is ./FM_index_Results}
}
\value{
A message of success and the location of the saved FM-Index files.
If an error occurs, an error message is returned, and the function
execution is stopped.
}
\description{
This function creates an FM-index for a given DNA sequence in FASTA format.
The FM-index is a data structure that is useful for an efficient pattern
search and storage. It consists of the \emph{Burrows-Wheeler Transform (BWT),
a sparse tally matrix, and a subset suffix array (SSA).}
The resulting data structures are saved as CSV files
in a specified directory.
This compressed version of the FM-Index can potentially store an entire human
genome in less than 2 GB
Increasing \code{step_tally} and \code{k}, leads to a more efficient storage,
on the other hand, it could lead to longer search times for patterns.
}
\examples{

file <- system.file("extdata", "prova.fasta", package = "FMIndex")
FM_index(file, step_tally = 10, k = 10)

}
\seealso{
\code{\link{createBWT}}, \code{\link{sparse_tally}},
\code{\link{create_SA}}
}
