% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMIndex_Script.R
\name{sparse_tally}
\alias{sparse_tally}
\title{\strong{Generate a Sparse Tally Matrix of a Burrows-Wheeler Transform}}
\usage{
sparse_tally(string, step)
}
\arguments{
\item{string}{A character string representing a DNA sequence, containing the
bases A, C, G, and T.
The BWT is generated within the function.}

\item{step}{An integer specifying the step size used for selecting rows from
the tally matrix to create its sparse version containing just the
selected rows}
}
\value{
A sparse tally matrix representing cumulative counts of each base
(A, C, G, T) in the BWT at the specified intervals.
The row names correspond to the positions in the original BWT.
}
\description{
This function generates a sparse tally matrix of the Burrows-Wheeler
Transform (BWT) of a given DNA string.
The BWT is generated inside the function, so the input is just the DNA string
and a step.
The sparse tally matrix is useful for efficiently storing cumulative
character counts at specific intervals, specified by the step parameter,
which can be used for further applications, such as FM-index creation.
}
\details{
The function first computes the BWT of the input string
using \link{createBWT}. It then calculates cumulative counts for
each unique character in the BWT.
The resulting sparse matrix stores these counts only at the
intervals defined by \code{step}, making it more memory-efficient while
retaining essential information for sequence analysis.
}
\examples{
DNA_string <- "ACGCGGATTGTGATC"
sparse_matrix <- sparse_tally(DNA_string, 2)
print(sparse_matrix)

}
\seealso{
\code{\link{createBWT}} for generating the Burrows-Wheeler Transform
}
